fix: stop exporting internal types
